project-{topic}-plus/backend/requirements.txt
```
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
sqlalchemy[asyncio]==2.0.23
asyncpg==0.28.0
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
aioredis==2.0.1 # For async Redis operations
fastapi-limiter==0.1.5 # For rate limiting
pydantic==2.5.2
pydantic-settings==2.1.0 # For FastAPI config management
httpx==0.25.2 # For API testing
pytest==7.4.3
pytest-asyncio==0.23.1
pytest-cov==4.1.0
flake8==6.1.0
black==23.11.0
isort==5.12.0
```